Add more parameters to G2Continuity, deprecate G3Continuity
We can do more with G2, achieving better results than G3 implementation

+++ capsule/src/main/java/com/kyant/capsule/G1Continuity.kt
@@ -36,8 +36,8 @@ data object G1Continuity : Continuity {
             is G1Continuity -> this
             is G2Continuity ->
                 G2Continuity(
-                    circleFraction = lerp(1f, stop.circleFraction, fraction),
-                    extendedFraction = lerp(0f, stop.extendedFraction, fraction)
+                    circleFraction = lerp(1f, stop.circleFraction.toFloat(), fraction).toDouble(),
+                    extendedFraction = lerp(0f, stop.extendedFraction.toFloat(), fraction).toDouble()
                 )
 
             is G3Continuity ->
